requirements/base-requirements.txt
Django==4.2.9
selenium==4.20.0
requests==2.31.0 
webdriver-manager==4.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-camel-case==1.4.2
django-cors-headers==4.3.1
django-split-settings==1.2.0
redis==5.0.1
django-redis==5.4.0
django-cachalot==2.6.1
django-extensions==3.2.3
drf-yasg==1.21.7
pydotplus==2.0.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
pillow==10.3.0 
